moving plugin logic from huntbot class to fluxrlplugin cog class
