+++ huntbot/HuntBot.py
@@ -39,17 +39,7 @@ class HuntBot:
         self.team_one_chat_channel_id = 0
         self.team_two_chat_channel_id = 0
         self.hunt_edition = ""
-        self.wom_competition_id = 0
         self.guild_id = 699971574689955850
-        self.participant_whitelist: set[str] = set()
-        self.monster_whitelist: set[str] = set()
-        self.item_whitelist: set[str] = set()
-        self.monster_whitelist_fp = "../conf/monster_whitelist.json"
-        self.item_whitelist_fp = "../conf/item_whitelist.json"
-
-        # Comp ID gets appended onto URLs later after config is retrieved
-        self.wom_event_api_url = "https://api.wiseoldman.net/v2/competitions/"
-        self.wom_event_website_url = "https://wiseoldman.net/competitions/"
 
         # TODO hardcode these for now
         self.general_channel_id = 699971574689955853
@@ -151,7 +141,6 @@ class HuntBot:
             self.team_one_chat_channel_id = int(self.config_map.get("TEAM_1_CHAT_CHANNEL_ID", "0"))
             self.team_two_chat_channel_id = int(self.config_map.get("TEAM_2_CHAT_CHANNEL_ID", "0"))
             self.hunt_edition = self.config_map.get("HUNT_EDITION", "")
-            self.wom_competition_id = self.config_map.get("WOM_COMPETITION_ID", "")
         except ValueError as e:
             logger.exception("Invalid type in config values (expected integer for channel IDs).", exc_info=e)
             raise InvalidConfig("Invalid type in config values: expected integers for channel IDs.")
@@ -180,15 +169,10 @@ class HuntBot:
             missing_fields.append("TEAM_2_CHAT_CHANNEL_ID")
         if not self.hunt_edition:
             missing_fields.append("HUNT_EDITION")
-        if not self.wom_competition_id:
-            missing_fields.append("WOM_COMPETITION_ID")
 
         if missing_fields:
             logger.error(f"Missing or invalid configuration fields: {', '.join(missing_fields)}")
             raise InvalidConfig(f"Missing or invalid configuration fields: {', '.join(missing_fields)}")
-
-        # Generate the WOM Competition URLs using the Comp ID from the configuration
-        self.generate_wom_competition_urls()
 
         # Combine the date and time strings
         start_datetime_str = f"{self.start_date} {self.start_time}"
@@ -241,55 +225,3 @@ class HuntBot:
         self.config_map['BOUNTY_PASSWORD'] = self.bounty_password
         self.config_map['DAILY_PASSWORD'] = self.daily_password
         self.config_map['ENDED'] = self.ended
-
-    def generate_participant_whitelist(self) -> None:
-        """
-        Iterates over the participations array in the JSON payload received from the WOM Hunt Competition query
-        and saves the whitelist in memory.
-        """
-        url = "https://api.wiseoldman.net/v2/competitions/100262"
-        r = requests.get(url)
-        data = r.json()
-
-        for player in data.get("participations", []):
-            player_name = player.get("player", {}).get("displayName", "")
-            if player_name:
-                self.participant_whitelist.add(player_name)
-
-    def print_whitelist(self) -> None:
-        print(sorted(self.participant_whitelist))
-
-    def generate_wom_competition_urls(self) -> None:
-        """
-        Appends the WOM Competition ID from the configuration to the end of the WOM event URls
-        """
-        self.wom_event_api_url = self.wom_event_api_url + str(self.wom_competition_id)
-        self.wom_event_website_url = self.wom_event_website_url + str(self.wom_competition_id)
-
-    def read_monster_whitelist_file(self) -> None:
-        with open(self.monster_whitelist_fp, "r", encoding="utf-8") as f:
-            monsters = json.load(f)
-
-        # Convert the list to a set
-        self.monster_whitelist = set(monsters)
-
-    def read_item_whitelist_file(self) -> None:
-        with open(self.item_whitelist_fp, "r", encoding="utf-8") as f:
-            items = json.load(f)
-
-        # Convert the list to a set
-        self.item_whitelist = set(items)
-
-    def generate_plugin_config_monster_list(self) -> None:
-        ...
-
-    def generate_plugin_config_item_list(self) -> None:
-        ...
-
-
-if __name__ == "__main__":
-    hunt_bot = HuntBot()
-    hunt_bot.read_monster_whitelist_file()
-    hunt_bot.read_item_whitelist_file()
-    print(hunt_bot.monster_whitelist)
-    print(hunt_bot.item_whitelist)
